Create separate files for sdv and kuksa API and make it changeable through protcol option
