+++ kuksa_databroker/databroker-cli/Cargo.toml
@@ -20,8 +20,8 @@ license = "Apache-2.0"
 
 [dependencies]
 common = { path = "../lib"}
-kuksa = { path = "../lib/kuksa", optional = true}
-sdv = { path = "../lib/sdv", optional = true}
+kuksa = { path = "../lib/kuksa"}
+sdv = { path = "../lib/sdv"}
 databroker-proto = { workspace = true }
 tonic = { workspace = true, features = ["transport", "channel", "prost"] }
 prost = { workspace = true }
@@ -44,7 +44,5 @@ regex = "1.6.0"
 http = "0.2.8"
 
 [features]
-default = ["tls", "feature_sdv"]
-feature_kuksa = ["kuksa"]
-feature_sdv = ["sdv"]
+default = ["tls"]
 tls = ["tonic/tls"]
